refactor: [YSS-66] 패키지 경로 변경
